Computes Predictions for the Exemplar-based Models (GCM, EBM)

Computes Predictions for the Exemplar-based Models (GCM, EBM)

Usage

ebm_cpp(
  criterion,
  features,
  w,
  r,
  q,
  lambda,
  b,
  wf,
  lastLearnTrial,
  firstOutTrial,
  init,
  has_criterion,
  similarity,
  ismultiplicative
)

Arguments

criterion

numeric vector with experienced criterion

features

numeric matrix with feature criterion

w

numeric vector of weights (model parameter)

r

order of Minkowski distance metic (model parameter)

q

relation between similarity and distance (model parameter)

lambda

sensitivity (model parameter)

b

bias parameter vector for classification (model parameter), must be NA for judgments

wf

weight vector with a weight for each feature combination

lastLearnTrial

integer last trial of learning phase

firstOutTrial

integer first trial of output, starting the predictions later

init

value for the initial trials

has_criterion

vector where a criterion is present

similarity

A string, the similarity function

ismultiplicative

A number (0 or 1), 1 means the combination of exemplars is multiplicative, i.e. multiplicative exemplar model

Examples

# none